requirements.txt
streamlit>=1.24.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
requests>=2.31.0